requirements.txt
streamlit
get-all-tickers
pandas
numpy
yfinance
seaborn
altair
stqdm
pillow